merge work (#4)
* create converable agent wrapepr for llamaindex aget

* fix autogen agent

* Add .devcontainer.json and readme.md files for each section of the project

- Added .devcontainer.json file with Python image and post-create command to install requirements.txt
- Added readme.md files for each section of the project, including agents_and_conversation, agents_with_tools, agents_that_learn, travel_agency_front_office, and travel_agency_with_front_and_back_office
- Renamed autogen/travel agency/requirements.txt to requirements.txt

* Add GitHub Codespaces badge to README.md

This commit adds a GitHub Codespaces badge to the README.md file. The badge allows users to easily open the project in GitHub Codespaces by clicking on it.
Add GitHub Codespaces badge to README.md

This commit adds a GitHub Codespaces badge to the README.md file. The badge allows users to easily open the project in GitHub Codespaces by clicking on it.

* Refactor tutorial links in README.md

- Updated tutorial links to be more descriptive
- Renamed tutorial links to reflect their content

* Update requirements.txt

- Updated duckdb to version 1.0.0
- Updated langchainhub to version 0.1.17
- Updated llama-index-core to version 0.10.43
- Updated llama-index-llms-openai to version 0.1.22
- Updated llama-index-postprocessor-rankgpt-rerank to version 0.1.6
- Updated llama-index-readers-file to version 0.1.23
- Updated llama-index-readers-web to version 0.1.18
- Updated pillow to version 10.3.0
- Updated pyautogen to version 0.2.28
- Updated PyMuPDF to version 1.24.

* feat: Add main.py for agents and conversation

This commit adds the `main.py` file which defines two agents, `commedian` and `user_proxy`. The `commedian` agent is configured as a professional commedian who can tell jokes and entertain people. The `user_proxy` agent is a user proxy that initiates the chat with a request for a joke about cats and ninjas.

* feat: Add customer assistant experience and improve customer service agent

- Added a new file `customer_assistant_experience/readme.md` to serve as a placeholder for persistent experience.
- Updated the `main.py` file in the travel agency with front and back office:
  - Improved the system message for the `customer_service` agent, providing instructions on accessing booking details and sending emails.
  - Added information about using customer booking habits to provide better service.
- Removed the `autogen/travel agency/readme.md` file.

This commit enhances the customer assistant experience and improves the functionality of the customer service agent.

* Add copyright and license information to main.py and llamaindex_conversable_agent.py

This commit adds the copyright notice and license information to the main.py and llamaindex_conversable_agent.py files. This ensures that the code is properly licensed under the MIT license.

* feat: Add Teachability capability to customer assistant

This commit adds the Teachability capability to the customer assistant. The `Teachability` class is imported from `autogen.agentchat.contrib.capabilities.teachability`. A new instance of `Teachability` is created and added to the `customer_assistant` agent. The database directory for storing customer assistant experiences is set to "./customer_assistant_experience".

* Rename READMES and doc directory

* Add name to devcontainer

* Add intro section docs

* Add more env vars and test 01 section sample

* Refactor Azure OpenAI embedding and LLM setup

- Updated the deployment name, API key, Azure endpoint, and API version for Azure OpenAI embedding and LLM
- Replaced references to `OpenAIEmbedding` with `AzureOpenAIEmbedding`
- Modified the configuration list for autogen to use Azure OpenAI parameters

The changes ensure that the code is compatible with the new Azure OpenAI models and configurations.

* Ignore SQLite3 files and a specific pickle file for customer assistant experience.

- Ignore *.sqlite3 files to prevent them from being tracked.
- Ignore the specific pickle file "uid_text_dict.pkl" in the "05_travel_agency_with_front_and_back_office/customer_assistant_experience" directory.

* Update requirements.txt

- Updated langchainhub to version 0.1.19
- Updated llama-index-core to version 0.10.44
- Updated llama-index-embeddings-azure-openai to version 0.1.10
- Updated llama-index-readers-file to version 0.1.25
- Updated llama-index to version 0.10.44

* feat: Add gitignore rules for pkl and bin files

- Added *.pkl and *.bin to the .gitignore file to exclude these files from version control.

feat: Add placeholder readme for comedian assistant experience

- Created a new readme.md file in the `03_agents_that_learn/comedian_assistant_experience` directory as a placeholder for persistent experience.

feat: Add main.py with code for comedian assistant

- Created a new main.py file in the `03_agents_that_learn` directory.
- Added code to initialize and configure the comedian assistant agent using Azure OpenAI API.
- Added code to set up teachability capabilities for the comedian assistant.
- Added code to initiate a chat between the user proxy agent and the comedian assistant, asking for a joke about cats and ninjas.

* Change reset_db flag in Teachability class from True to False

Teachability: Change reset_db flag to False

* Update max_iterations parameter in location_specialist initialization

The max_iterations parameter in the initialization of the location_specialist object has been updated from 30 to 5. This change was made to limit the number of iterations performed by the agent.

* Refactor LLamaIndexConversableAgent and add Teachability to trip_assistant

- Refactored the `LLamaIndexConversableAgent` class in `llamaindex_conversable_agent.py` by reformatting the code, adding type hints, and improving code readability.
- Added a new instance of `Teachability` called `trip_assistant_experience` to the `trip_assistant` agent in `main.py`.
- The purpose of this change is to enhance the trip assistant's ability to learn from customer interactions and improve its responses over time.

* feat: Add ChatSummaryMemoryBuffer for improved memory management

This commit adds the `ChatSummaryMemoryBuffer` class to the codebase. This class is used to manage chat summaries and improve memory usage. It allows for better tracking of customer interactions and helps in providing more personalized service based on customer interests and preferences. The `max_iterations` parameter has also been increased to 8 for the `location_specialist` agent, allowing for a more thorough search process.

* Update

* Update Env Gitignore + DevContainer Setup

* Update

---------

Co-authored-by: Luis Quintanilla <46974588+luisquintanilla@users.noreply.github.com>
Co-authored-by: Lee Stott <leestott@microsoft.com>