Add support for Autogen framework in conversational agent cookbooks
The commit adds support for the Autogen framework in the conversational agent cookbooks. Now, all the samples in the cookbooks will be using the Autogen framework. This change aims to enhance the functionality and capabilities of the conversational agents by leveraging the features provided by Autogen.

Areas tested: Conversational agent cookbooks with Autogen framework